wrapper functions fix

+++ exchanges/deribit/deribit_types.go
@@ -201,7 +201,7 @@ type InstrumentData struct {
 		TickSize   float64 `json:"tick_size"`
 	} `json:"tick_size_steps"`
 	BlockTradeTickSize       float64 `json:"block_trade_tick_size"`
-	BlockTradeMinTradeAmount int64   `json:"block_trade_min_trade_amount"`
+	BlockTradeMinTradeAmount float64 `json:"block_trade_min_trade_amount"`
 	InstrumentType           string  `json:"instrument_type"`
 }
